chore: 🤖 more annotation comments gen
